first part
The first part of the project was developed by me, Abdessamad Mekkaoui. In this section, I created and installed the necessary packages and designed the database architecture. I then linked both the Flask app and the database. Afterward, I developed the React app, which includes a form where the admin can add new students to the database.